Remove `-recipe` from recipe ids
The Factoriopedia in 2.0 merges item, entity, and recipe,
(and fluid and tile) entries that have identical ids.
As such, ending recipe ids with `-recipe` has an adverse
effect on players using the Factoriopedia to help plan their
factories. This change results in many seemingly duplicate
Factoriopedia entries getting merged into single entries.

This change is sweeping and likely has some spots that
were changed eroneously or were missed, but I could not
find them when browsing the diff. An in-game check reveals
that some entries are still split between item and recipe
despite now having identical ids, though I'm not sure why.
Unpasturised milk has `unpasturized-milk` for the fluid,
but `u-milk` for the recipes, and this change does not touch
this, so they remain duplicated.

I also did not add migrations for the change largely because
I've never dealt with them before and I'm less confident
in my ability to add them.

This change is incompatible with the current version of
Baketorio+, which directly references some Baketorio
recipes by id. It *might* be possible to omit those
specific changes if compatibility is desired, but a better
solution would be for that mod to open a repo so a
corresponding patch could be sent there as well.